+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraConnectionIntegrationTest.java
@@ -39,8 +39,6 @@ public class CassandraConnectionIntegrationTest {
 
     @Test
     public void testAuthMissing() {
-        CassandraKeyValueServiceImpl.createForTesting(
-                NO_CREDS_CKVS_CONFIG,
-                CassandraResource.LEADER_CONFIG).close();
+        CassandraKeyValueServiceImpl.createForTesting(NO_CREDS_CKVS_CONFIG).close();
     }
 }