+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraKeyValueServiceImpl.java
@@ -30,6 +30,8 @@ import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import java.util.stream.Collectors;
 
+import javax.annotation.Nullable;
+
 import org.apache.cassandra.thrift.CASResult;
 import org.apache.cassandra.thrift.CfDef;
 import org.apache.cassandra.thrift.Column;
@@ -70,7 +72,6 @@ import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfig;
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceRuntimeConfig;
 import com.palantir.atlasdb.cassandra.CassandraMutationTimestampProvider;
 import com.palantir.atlasdb.cassandra.CassandraMutationTimestampProviders;
-import com.palantir.atlasdb.config.LeaderConfig;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.BatchColumnRangeSelection;
 import com.palantir.atlasdb.keyvalue.api.CandidateCellForSweeping;
@@ -201,8 +202,6 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     private final CassandraKeyValueServiceConfig config;
     private final CassandraClientPool clientPool;
 
-    private final Optional<LeaderConfig> leaderConfig;
-
     private ConsistencyLevel readConsistency = ConsistencyLevel.LOCAL_QUORUM;
 
     private final TracingQueryRunner queryRunner;
@@ -223,15 +222,13 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
 
     private final CassandraMutationTimestampProvider mutationTimestampProvider;
 
-    public static CassandraKeyValueService createForTesting(
-            CassandraKeyValueServiceConfig config,
-            Optional<LeaderConfig> leaderConfig) {
+    public static CassandraKeyValueService createForTesting(CassandraKeyValueServiceConfig config) {
         MetricsManager metricsManager = MetricsManagers.createForTests();
         CassandraClientPool clientPool = CassandraClientPoolImpl.createImplForTest(metricsManager,
                 config,
                 CassandraClientPoolImpl.StartupChecks.RUN,
                 new Blacklist(config));
-        return createOrShutdownClientPool(metricsManager, config, clientPool, leaderConfig,
+        return createOrShutdownClientPool(metricsManager, config, clientPool,
                 CassandraMutationTimestampProviders.legacyModeForTestsOnly(),
                 LoggerFactory.getLogger(CassandraKeyValueService.class),
                 AtlasDbConstants.DEFAULT_INITIALIZE_ASYNC);
@@ -240,13 +237,11 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     public static CassandraKeyValueService create(
             MetricsManager metricsManager,
             CassandraKeyValueServiceConfig config,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider) {
         return create(
                 metricsManager,
                 config,
                 CassandraKeyValueServiceRuntimeConfig::getDefault,
-                leaderConfig,
                 mutationTimestampProvider,
                 AtlasDbConstants.DEFAULT_INITIALIZE_ASYNC);
     }
@@ -254,13 +249,11 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     public static CassandraKeyValueService create(
             MetricsManager metricsManager,
             CassandraKeyValueServiceConfig config,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider,
             CassandraClientPool clientPool) {
         return createOrShutdownClientPool(metricsManager,
                 config,
                 clientPool,
-                leaderConfig,
                 mutationTimestampProvider,
                 LoggerFactory.getLogger(CassandraKeyValueService.class),
                 AtlasDbConstants.DEFAULT_INITIALIZE_ASYNC);
@@ -270,13 +263,11 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
             MetricsManager metricsManager,
             CassandraKeyValueServiceConfig config,
             java.util.function.Supplier<CassandraKeyValueServiceRuntimeConfig> runtimeConfig,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider,
             boolean initializeAsync) {
         return create(metricsManager,
                 config,
                 runtimeConfig,
-                leaderConfig,
                 mutationTimestampProvider,
                 LoggerFactory.getLogger(CassandraKeyValueService.class),
                 initializeAsync);
@@ -286,13 +277,11 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     static CassandraKeyValueService create(
             MetricsManager metricsManager,
             CassandraKeyValueServiceConfig config,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider,
             Logger log) {
         return create(metricsManager,
                 config,
                 CassandraKeyValueServiceRuntimeConfig::getDefault,
-                leaderConfig,
                 mutationTimestampProvider,
                 log,
                 AtlasDbConstants.DEFAULT_INITIALIZE_ASYNC);
@@ -302,7 +291,6 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
             MetricsManager metricsManager,
             CassandraKeyValueServiceConfig config,
             java.util.function.Supplier<CassandraKeyValueServiceRuntimeConfig> runtimeConfig,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider,
             Logger log,
             boolean initializeAsync) {
@@ -310,7 +298,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
                 config,
                 runtimeConfig,
                 initializeAsync);
-        return createOrShutdownClientPool(metricsManager, config, clientPool, leaderConfig, mutationTimestampProvider,
+        return createOrShutdownClientPool(metricsManager, config, clientPool, mutationTimestampProvider,
                 log, initializeAsync);
     }
 
@@ -318,12 +306,11 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
             MetricsManager metricsManager,
             CassandraKeyValueServiceConfig config,
             CassandraClientPool clientPool,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider,
             Logger log,
             boolean initializeAsync) {
         try {
-            return createAndInitialize(metricsManager, config, clientPool, leaderConfig,
+            return createAndInitialize(metricsManager, config, clientPool,
                     mutationTimestampProvider, log, initializeAsync);
         } catch (Exception e) {
             log.warn("Error occurred in creating Cassandra KVS. Now attempting to shut down client pool...", e);
@@ -342,7 +329,6 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
             MetricsManager metricsManager,
             CassandraKeyValueServiceConfig config,
             CassandraClientPool clientPool,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider,
             Logger log,
             boolean initializeAsync) {
@@ -351,7 +337,6 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
                 metricsManager,
                 config,
                 clientPool,
-                leaderConfig,
                 mutationTimestampProvider);
         keyValueService.wrapper.initialize(initializeAsync);
         return keyValueService.wrapper.isInitialized() ? keyValueService : keyValueService.wrapper;
@@ -360,14 +345,12 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     private CassandraKeyValueServiceImpl(Logger log,
             MetricsManager metricsManager, CassandraKeyValueServiceConfig config,
             CassandraClientPool clientPool,
-            Optional<LeaderConfig> leaderConfig,
             CassandraMutationTimestampProvider mutationTimestampProvider) {
         super(createInstrumentedFixedThreadPool(config, metricsManager.getRegistry()));
         this.log = log;
         this.metricsManager = metricsManager;
         this.config = config;
         this.clientPool = clientPool;
-        this.leaderConfig = leaderConfig;
         this.mutationTimestampProvider = mutationTimestampProvider;
         this.queryRunner = new TracingQueryRunner(log, tracingPrefs);
         this.wrappingQueryRunner = new WrappingQueryRunner(queryRunner);
@@ -464,7 +447,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         }
     }
 
-    private Optional<byte[]> lookupClusterSideMetadata(
+    private static Optional<byte[]> lookupClusterSideMetadata(
             Map<TableReference, byte[]> metadataForTables,
             TableReference tableRef) {
         // TODO (jkong): Replace with or() once we can use Java 9
@@ -600,7 +583,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         }
     }
 
-    private List<ByteBuffer> wrap(List<byte[]> arrays) {
+    private static List<ByteBuffer> wrap(List<byte[]> arrays) {
         List<ByteBuffer> byteBuffers = Lists.newArrayListWithCapacity(arrays.size());
         for (byte[] r : arrays) {
             byteBuffers.add(ByteBuffer.wrap(r));
@@ -656,7 +639,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
             }
 
             SetMultimap<Long, Cell> cellsByTs = Multimaps.invertFrom(
-                    Multimaps.forMap(timestampByCell), HashMultimap.<Long, Cell>create());
+                    Multimaps.forMap(timestampByCell), HashMultimap.create());
             Builder<Cell, Value> builder = ImmutableMap.builder();
             for (long ts : cellsByTs.keySet()) {
                 StartTsResultsCollector collector = new StartTsResultsCollector(metricsManager, ts);
@@ -894,7 +877,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         }.iterator());
     }
 
-    private boolean isEndOfColumnRange(boolean completedCell, byte[] lastCol, int numRawResults,
+    private static boolean isEndOfColumnRange(boolean completedCell, byte[] lastCol, int numRawResults,
                                        BatchColumnRangeSelection columnRangeSelection) {
         return (numRawResults < columnRangeSelection.getBatchHint())
                 || (completedCell
@@ -904,7 +887,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
                             columnRangeSelection.getEndCol())));
     }
 
-    private byte[] getNextColumnRangeColumn(boolean completedCell, byte[] lastCol) {
+    private static byte[] getNextColumnRangeColumn(boolean completedCell, byte[] lastCol) {
         if (!completedCell) {
             return lastCol;
         } else {
@@ -912,7 +895,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         }
     }
 
-    private Range createColumnRange(byte[] startColOrEmpty, byte[] endColExlusiveOrEmpty, long startTs) {
+    private static Range createColumnRange(byte[] startColOrEmpty, byte[] endColExlusiveOrEmpty, long startTs) {
         ByteBuffer start = startColOrEmpty.length == 0
                 ? Range.UNBOUND_START
                 : Range.startOfColumn(startColOrEmpty, startTs);
@@ -1025,7 +1008,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         return tasks;
     }
 
-    private Set<TableReference> extractTableNames(Iterable<TableCellAndValue> tableCellAndValues) {
+    private static Set<TableReference> extractTableNames(Iterable<TableCellAndValue> tableCellAndValues) {
         Set<TableReference> tableRefs = Sets.newHashSet();
         for (TableCellAndValue tableCellAndValue : tableCellAndValues) {
             tableRefs.add(tableCellAndValue.tableRef);
@@ -1052,7 +1035,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         });
     }
 
-    private MutationMap convertToMutations(List<TableCellAndValue> batch, long timestamp) {
+    private static MutationMap convertToMutations(List<TableCellAndValue> batch, long timestamp) {
         MutationMap mutationMap = new MutationMap();
         for (TableCellAndValue tableCellAndValue : batch) {
             Cell cell = tableCellAndValue.cell;
@@ -1386,14 +1369,11 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         return Optional.ofNullable(getMetadataForTables().get(tableRef)).orElse(AtlasDbConstants.EMPTY_TABLE_METADATA);
     }
 
-    private boolean matchingIgnoreCase(TableReference t1, TableReference t2) {
+    private static boolean matchingIgnoreCase(@Nullable TableReference t1, TableReference t2) {
         if (t1 != null) {
             return t1.getQualifiedName().toLowerCase().equals(t2.getQualifiedName().toLowerCase());
         } else {
-            if (t2 == null) {
-                return true;
-            }
-            return false;
+            return t2 == null;
         }
     }
 
@@ -1497,7 +1477,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         }
     }
 
-    private boolean metadataIsDifferent(byte[] existingMetadata, byte[] requestMetadata) {
+    private static boolean metadataIsDifferent(byte[] existingMetadata, byte[] requestMetadata) {
         return !Arrays.equals(existingMetadata, requestMetadata);
     }
 
@@ -1524,7 +1504,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         }
     }
 
-    private String schemaChangeDescriptionForPutMetadataForTables(Collection<CfDef> updatedCfs) {
+    private static String schemaChangeDescriptionForPutMetadataForTables(Collection<CfDef> updatedCfs) {
         String tables = updatedCfs.stream()
                 .map(CassandraKeyValueServices::tableReferenceFromCfDef)
                 .map(Object::toString)
@@ -1768,7 +1748,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         return clusterStatus;
     }
 
-    private boolean isClusterQuorumAvaialble(ClusterAvailabilityStatus clusterStatus) {
+    private static boolean isClusterQuorumAvaialble(ClusterAvailabilityStatus clusterStatus) {
         return clusterStatus.equals(ClusterAvailabilityStatus.ALL_AVAILABLE)
                 || clusterStatus.equals(ClusterAvailabilityStatus.QUORUM_AVAILABLE);
     }