+++ atlasdb-container-test-utils/src/main/java/com/palantir/atlasdb/containers/UninitializedCassandraResource.java
@@ -72,7 +72,6 @@ public class UninitializedCassandraResource extends ExternalResource {
                 MetricsManagers.createForTests(),
                 containerInstance.getConfig(),
                 CassandraKeyValueServiceRuntimeConfig::getDefault,
-                CassandraContainer.LEADER_CONFIG,
                 CassandraMutationTimestampProviders.legacyModeForTestsOnly(),
                 true);
     }