+++ leader-election-impl/src/main/java/com/palantir/paxos/PaxosResponses.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,26 +13,74 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.palantir.paxos;
 
-import com.google.common.base.Predicate;
-import com.palantir.paxos.persistence.generated.remoting.PaxosAcceptorPersistence;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Stream;
 
-public final class PaxosResponses {
-    private PaxosResponses() {}
+public class PaxosResponses<T extends PaxosResponse> {
+    private final int totalRequests;
+    private final int quorum;
+    private final boolean shortcut;
+    private List<T> responses = new ArrayList<>();
+    private int successes = 0;
+    private int failures = 0;
 
-    public static Predicate<PaxosResponse> isSuccessfulPredicate() {
-        return response -> response != null && response.isSuccessful();
+    PaxosResponses(int totalRequests, int quorum, boolean shortcut) {
+        this.totalRequests = totalRequests;
+        this.quorum = quorum;
+        this.shortcut = shortcut;
     }
 
-    public static PaxosAcceptorPersistence.PaxosResponse toProto(PaxosResponse result) {
-        return PaxosAcceptorPersistence.PaxosResponse.newBuilder()
-                .setAck(result.isSuccessful())
-                .build();
+    public boolean hasMoreRequests() {
+        return successes + failures < totalRequests;
     }
 
-    public static PaxosResponse fromProto(PaxosAcceptorPersistence.PaxosResponse proto) {
-        boolean ack = proto.getAck();
-        return new PaxosResponseImpl(ack);
+    boolean shouldProcessNextRequest() {
+        return !hasQuorum() && !shouldGiveUpOnAchievingQuorum();
+    }
+
+    public void add(T response) {
+        if (response.isSuccessful()) {
+            successes++;
+        } else {
+            failures++;
+        }
+        responses.add(response);
+    }
+
+    void markFailure() {
+        failures++;
+    }
+
+    public List<T> get() {
+        return responses;
+    }
+
+    public Stream<T> stream() {
+        return responses.stream();
+    }
+
+    public boolean hasQuorum() {
+        return successes >= quorum;
+    }
+
+    PaxosQuorumStatus getQuorumResult() {
+        if (hasQuorum()) {
+            return PaxosQuorumStatus.QUORUM_AGREED;
+        } else if (thereWereDisagreements()) {
+            return PaxosQuorumStatus.SOME_DISAGREED;
+        }
+        return PaxosQuorumStatus.NO_QUORUM;
+    }
+
+    private boolean shouldGiveUpOnAchievingQuorum() {
+        return shortcut && failures > totalRequests - quorum;
+    }
+
+    private boolean thereWereDisagreements() {
+        return responses.stream().anyMatch(response -> !response.isSuccessful());
     }
 }