+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/paxos/PaxosTimestampBoundStore.java
@@ -30,6 +30,7 @@ import org.slf4j.LoggerFactory;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Iterables;
 import com.google.common.collect.Ordering;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.common.concurrent.PTExecutors;
@@ -43,6 +44,7 @@ import com.palantir.paxos.PaxosQuorumChecker;
 import com.palantir.paxos.PaxosResponse;
 import com.palantir.paxos.PaxosRoundFailureException;
 import com.palantir.paxos.PaxosValue;
+import com.palantir.paxos.PaxosResponses;
 import com.palantir.timestamp.DebugLogger;
 import com.palantir.timestamp.MultipleRunningTimestampServiceError;
 import com.palantir.timestamp.TimestampBoundStore;
@@ -51,7 +53,6 @@ public class PaxosTimestampBoundStore implements TimestampBoundStore {
     private static final Logger log = LoggerFactory.getLogger(PaxosTimestampBoundStore.class);
 
     private static final int QUORUM_OF_ONE = 1;
-    private static final boolean ONLY_LOG_ON_QUORUM_FAILURE = true;
 
     private final PaxosProposer proposer;
     private final PaxosLearner knowledge;
@@ -106,17 +107,16 @@ public class PaxosTimestampBoundStore implements TimestampBoundStore {
      * @throws ServiceNotAvailableException if we couldn't contact a quorum
      */
     private List<PaxosLong> getLatestSequenceNumbersFromAcceptors() {
-        List<PaxosLong> responses = PaxosQuorumChecker.collectQuorumResponses(
+        PaxosResponses<PaxosLong> responses = PaxosQuorumChecker.collectQuorumResponses(
                 ImmutableList.copyOf(acceptors),
                 acceptor -> ImmutablePaxosLong.of(acceptor.getLatestSequencePreparedOrAccepted()),
                 proposer.getQuorumSize(),
                 executor,
-                PaxosQuorumChecker.DEFAULT_REMOTE_REQUESTS_TIMEOUT_IN_SECONDS,
-                ONLY_LOG_ON_QUORUM_FAILURE);
-        if (!PaxosQuorumChecker.hasQuorum(responses, proposer.getQuorumSize())) {
+                PaxosQuorumChecker.DEFAULT_REMOTE_REQUESTS_TIMEOUT);
+        if (!responses.hasQuorum()) {
             throw new ServiceNotAvailableException("could not get a quorum");
         }
-        return responses;
+        return responses.get();
     }
 
     /**
@@ -208,17 +208,15 @@ public class PaxosTimestampBoundStore implements TimestampBoundStore {
         if (seq <= PaxosAcceptor.NO_LOG_ENTRY) {
             return Optional.of(ImmutableSequenceAndBound.of(PaxosAcceptor.NO_LOG_ENTRY, 0L));
         }
-        List<PaxosLong> responses = PaxosQuorumChecker.collectQuorumResponses(
+        PaxosResponses<PaxosLong> responses = PaxosQuorumChecker.collectQuorumResponses(
                 ImmutableList.copyOf(learners),
                 learner -> getLearnedValue(seq, learner),
                 QUORUM_OF_ONE,
                 executor,
-                PaxosQuorumChecker.DEFAULT_REMOTE_REQUESTS_TIMEOUT_IN_SECONDS,
-                ONLY_LOG_ON_QUORUM_FAILURE);
-        if (responses.isEmpty()) {
-            return Optional.empty();
-        }
-        return Optional.of(ImmutableSequenceAndBound.of(seq, responses.iterator().next().getValue()));
+                PaxosQuorumChecker.DEFAULT_REMOTE_REQUESTS_TIMEOUT);
+        return Optional.ofNullable(Iterables.getFirst(responses.get(), null))
+                .map(PaxosLong::getValue)
+                .map(value -> ImmutableSequenceAndBound.of(seq, value));
     }
 
     /**